# swagger.yaml
openapi: 3.0.0
info:
  title: Recipe API
  description: Backend API for managing and sharing recipes - ITIS 4166-051 Group 17
  version: 1.0.0
  contact:
    name: Group 17
servers:
  - url: https://recipe-api-qkaa.onrender.com/api
    description: Production server (Render)
  - url: http://localhost:1234/api
    description: Local server (primary)
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Unauthorized - Authentication required
    Forbidden:
      description: Forbidden - Insufficient permissions
...
    Success:
      description: Operation successful
...
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Get all users (Admin only)
      security: [{ BearerAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /recipes:
    get:
      tags: [Recipes]
      summary: Get all recipes
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Recipes]
      summary: Create a new recipe
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, prepTime, cookTime, servings]
              properties:
                title: { type: string, minLength: 3, maxLength: 100 }
                description: { type: string, maxLength: 500 }
                prepTime: { type: integer, minimum: 0 }
                cookTime: { type: integer, minimum: 0 }
                servings: { type: integer, minimum: 1 }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /ingredients:
    get:
      tags: [Ingredients]
      summary: Get all ingredients
      responses:
        '200': { $ref: '#/components/responses/Success' }
    post:
      tags: [Ingredients]
      summary: Create ingredient (Admin only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, unit]
              properties:
                name: { type: string, minLength: 2, maxLength: 100 }
                unit: { type: string, minLength: 1, maxLength: 20 }
      responses:
        '201': { $ref: '#/components/responses/Success' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...